refactor: use `MinMax` interface.
