+++ src/lib/number-between.type.ts
@@ -1,7 +1,7 @@
 // Type.
 import { AnyNumber } from './any-number.type';
 // Interface.
-import { MinMax } from '../interface/min-max.interface';
+import { MinMax } from './min-max.interface';
 /**
  * A `number` type or an instance of `Number` between a specified range takes generic type variable `Min` and `Max` constrained by `number`
  * type as range, and takes generic type variable `Type` constrained by generic type `AnyNumber` as the type.
